added tests to new stacks
